trying to fix GPU usage

+++ lstm_tf_imdb_modified.py
@@ -83,7 +83,7 @@ class LSTM_Model(object):
         # learning rate as a tf variable. Its value is therefore session dependent
         self._lr = tf.Variable(config.learning_rate, trainable=False)
         with tf.device("/cpu:0"):
-            self._inputs = tf.placeholder(tf.float32,[MAXLEN,BATCH_SIZE],name='embedded_inputs')
+            self._inputs = tf.placeholder(tf.int64,[MAXLEN,BATCH_SIZE],name='embedded_inputs')
             self._targets = tf.placeholder(tf.float32, [None, 2],name='targets')
             self._mask = tf.placeholder(tf.float32, [None, None],name='mask')
 
@@ -101,7 +101,7 @@ class LSTM_Model(object):
                 word_embedding = tf.get_variable('word_embedding', shape=[vocabulary_size, dim_proj],
                                                   initializer=tf.constant_initializer(random_embedding),dtype=tf.float32)
 
-                unrolled_inputs=tf.reshape(self._inputs,[1,-1],dtype=tf.int64)
+                unrolled_inputs=tf.reshape(self._inputs,[1,-1])
                 embedded_inputs = tf.nn.embedding_lookup(word_embedding, unrolled_inputs)
                 embedded_inputs = tf.reshape(embedded_inputs, [MAXLEN, BATCH_SIZE, dim_proj])
